Implements both malloc() and free()
I added a linked list implementation to store available memory blocks in a nodes. Free also checks to see if there are any subsequent blocks of memory and merges them if that is the case. Currently the runtime of this is O(n) but could be improved by sorting the list and running a binary search on it.
For my implementation of malloc I noticed the comment suggesting to break blocks much larger than n into smaller blocks and add the new smaller blocks to the list, but this seemed counter productive as free() would re-merge those blocks since they have subsequent addresses. Instead of the block is larger than size n, n's amount of memory is given to the user and the remaining memory in that block is reconfigured into a new node that sits in the same position that it held before in the linked list.